requirements.txt
attrs
pytest
pytest-cov
flake8